Update supervisord.conf
(a) In supervisord.conf you run ovsdb-server with the cmd line switch "--remote=ptcp:6640".

(b) Then in configure-ovs.sh you run "ovs-vsctl --no-wait -- set-manager ptcp:6640"

What happens is that at point (a) ovsdb-server creates the manager entry and listens and receives connections on port 6640. This happens before the various initializations in configure-ovs.sh. In addition to this, when (b) finally occurs ovsdb-server closes any connections it had received.

As you may guess, all this happens really, really quickly but it does occur in automated testing. Our controller was connects (sometimes) between (a) and (b) and "interesting" stuff ensues. 

I've tested this change on my setup and it works just fine.